requirements.txt
pytest==7.1.2
requests==2.27.1
selenium==4.2.0
allure-pytest==2.9.45
pytest-xdist==2.5.0
